+++ src/test/java/org/springframework/integration/kafka/listener/MultiBroker5ReplicatedTest.java
@@ -26,10 +26,10 @@ import org.springframework.integration.kafka.rule.KafkaEmbedded;
 /**
  * @author Marius Bogoevici
  */
-public class TestSingleBrokerWithCompression extends AbstractMessageListenerContainerTest {
+public class MultiBroker5ReplicatedTest extends AbstractMessageListenerContainerTest {
 
 	@Rule
-	public final KafkaEmbedded kafkaEmbeddedBrokerRule = new KafkaEmbedded(1);
+	public KafkaEmbedded kafkaEmbeddedBrokerRule = new KafkaEmbedded(5);
 
 	@Override
 	public KafkaEmbedded getKafkaRule() {
@@ -38,56 +38,57 @@ public class TestSingleBrokerWithCompression extends AbstractMessageListenerCont
 
 	@Test
 	public void testLowVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 2, 5, 100, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 3);
+		runMessageListenerTest(100, 2, 5, 100, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 2, 5, 1000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 3);
+		runMessageListenerTest(100, 2, 5, 1000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test @Ignore
 	public void testHighVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 2, 5, 10000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 3);
+		runMessageListenerTest(100, 2, 5, 10000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testLowVolumeMediumConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 5, 5, 100, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 3);
+		runMessageListenerTest(100, 5, 5, 100, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeMediumConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 5, 5, 1000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 3);
+		runMessageListenerTest(100, 5, 5, 1000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test @Ignore
 	public void testHighVolumeMediumConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 1, 1);
-		runMessageListenerTest(10000, 5, 5, 100000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 5, 5, 1);
+		runMessageListenerTest(100, 5, 5, 100000, 1, 0, TEST_TOPIC);
 	}
+
 
 	@Test
 	public void testLowVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 1, 1);
-		runMessageListenerTest(10000, 20, 100, 1000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 100, 5, 3);
+		runMessageListenerTest(100, 20, 100, 1000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 1, 1);
-		runMessageListenerTest(10000, 20, 100, 10000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 100, 5, 3);
+		runMessageListenerTest(100, 20, 100, 10000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test @Ignore
 	public void testHighVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 1, 1);
-		runMessageListenerTest(10000, 20, 100, 100000, 1, 1, TEST_TOPIC);
+		createTopic(TEST_TOPIC, 100, 5, 3);
+		runMessageListenerTest(100, 20, 100, 100000, 1, 0, TEST_TOPIC);
 	}
