+++ src/test/java/org/springframework/integration/kafka/listener/MultiBroker5ReplicatedWithBounceTest.java
@@ -34,48 +34,42 @@ import kafka.utils.VerifiableProperties;
 import org.junit.Rule;
 import org.junit.Test;
 
-import org.springframework.integration.kafka.core.Configuration;
 import org.springframework.integration.kafka.core.ConnectionFactory;
 import org.springframework.integration.kafka.core.KafkaMessage;
 import org.springframework.integration.kafka.core.Partition;
-import org.springframework.integration.kafka.core.ZookeeperConfiguration;
 import org.springframework.integration.kafka.rule.KafkaEmbedded;
-import org.springframework.integration.kafka.support.ZookeeperConnect;
 
 /**
  * @author Marius Bogoevici
  */
-public class TestZookeeperConfiguration extends AbstractMessageListenerContainerTest {
+
+public class MultiBroker5ReplicatedWithBounceTest extends AbstractMessageListenerContainerTest {
 
 	@Rule
-	public KafkaEmbedded kafkaEmbeddedBrokerRule = new KafkaEmbedded(2);
+	public KafkaEmbedded kafkaEmbeddedBrokerRule = new KafkaEmbedded(5);
 
 	@Override
 	public KafkaEmbedded getKafkaRule() {
 		return kafkaEmbeddedBrokerRule;
 	}
 
-	@Override
-	public Configuration getKafkaConfiguration() {
-		ZookeeperConnect zookeeperConnect = new ZookeeperConnect();
-		zookeeperConnect.setZkConnect(kafkaEmbeddedBrokerRule.getZookeeperConnectionString());
-		return new ZookeeperConfiguration(zookeeperConnect);
-	}
-
 	@Test
 	public void testLowVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 2, 1);
+		createTopic(TEST_TOPIC, 5, 5, 3);
 
 		ConnectionFactory connectionFactory = getKafkaBrokerConnectionFactory();
 		ArrayList<Partition> readPartitions = new ArrayList<Partition>();
 		for (int i = 0; i < 5; i++) {
-			readPartitions.add(new Partition(TEST_TOPIC, i));
+				readPartitions.add(new Partition(TEST_TOPIC, i));
 		}
 		final KafkaMessageListenerContainer kafkaMessageListenerContainer = new KafkaMessageListenerContainer(connectionFactory, readPartitions.toArray(new Partition[readPartitions.size()]));
 		kafkaMessageListenerContainer.setMaxFetch(100);
 		kafkaMessageListenerContainer.setConcurrency(2);
 
-		int expectedMessageCount = 100;
+		final int expectedMessageCount = 100;
+
+		createStringProducer(0).send(createMessages(100, TEST_TOPIC));
+
 
 		final MutableListMultimap<Integer,KeyedMessageWithOffset> receivedData = new SynchronizedPutFastListMultimap<Integer, KeyedMessageWithOffset>();
 		final CountDownLatch latch = new CountDownLatch(expectedMessageCount);
@@ -90,9 +84,13 @@ public class TestZookeeperConfiguration extends AbstractMessageListenerContainer
 
 		kafkaMessageListenerContainer.start();
 
-		createStringProducer(0).send(createMessages(100, TEST_TOPIC));
+		for (int i = 0; i < 5; i++) {
+			if (i % 2 == 0) {
+				kafkaEmbeddedBrokerRule.bounce(i);
+			}
+		}
 
-		latch.await((expectedMessageCount/5000) + 1, TimeUnit.MINUTES);
+		latch.await(50, TimeUnit.SECONDS);
 		kafkaMessageListenerContainer.stop();
 
 		assertThat(receivedData.valuesView().toList(), hasSize(expectedMessageCount));
@@ -102,6 +100,5 @@ public class TestZookeeperConfiguration extends AbstractMessageListenerContainer
 		validateMessageReceipt(receivedData, 2, 5, 100, expectedMessageCount, readPartitions, 1);
 
 	}
-
 
 }
