+++ src/test/java/org/springframework/integration/kafka/listener/MultiBroker5Test.java
@@ -26,7 +26,7 @@ import org.springframework.integration.kafka.rule.KafkaEmbedded;
 /**
  * @author Marius Bogoevici
  */
-public class TestMultiBroker5Replicated extends AbstractMessageListenerContainerTest {
+public class MultiBroker5Test extends AbstractMessageListenerContainerTest {
 
 	@Rule
 	public KafkaEmbedded kafkaEmbeddedBrokerRule = new KafkaEmbedded(5);
@@ -38,35 +38,37 @@ public class TestMultiBroker5Replicated extends AbstractMessageListenerContainer
 
 	@Test
 	public void testLowVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 5, 3);
+		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 2, 5, 100, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 5, 3);
+		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 2, 5, 1000, 1, 0, TEST_TOPIC);
 	}
 
-	@Test @Ignore
+	@Test
+	@Ignore
 	public void testHighVolumeLowConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 5, 3);
+		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 2, 5, 10000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testLowVolumeMediumConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 5, 3);
+		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 5, 5, 100, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeMediumConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 5, 5, 3);
+		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 5, 5, 1000, 1, 0, TEST_TOPIC);
 	}
 
-	@Test @Ignore
+	@Test
+	@Ignore
 	public void testHighVolumeMediumConcurrency() throws Exception {
 		createTopic(TEST_TOPIC, 5, 5, 1);
 		runMessageListenerTest(100, 5, 5, 100000, 1, 0, TEST_TOPIC);
@@ -75,21 +77,21 @@ public class TestMultiBroker5Replicated extends AbstractMessageListenerContainer
 
 	@Test
 	public void testLowVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 5, 3);
+		createTopic(TEST_TOPIC, 100, 5, 1);
 		runMessageListenerTest(100, 20, 100, 1000, 1, 0, TEST_TOPIC);
 	}
 
 	@Test
 	public void testMediumVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 5, 3);
+		createTopic(TEST_TOPIC, 100, 5, 1);
 		runMessageListenerTest(100, 20, 100, 10000, 1, 0, TEST_TOPIC);
 	}
 
-	@Test @Ignore
+	@Test
+	@Ignore
 	public void testHighVolumeHighConcurrency() throws Exception {
-		createTopic(TEST_TOPIC, 100, 5, 3);
+		createTopic(TEST_TOPIC, 100, 5, 1);
 		runMessageListenerTest(100, 20, 100, 100000, 1, 0, TEST_TOPIC);
 	}
-
 
 }
